+++ automation/src/issues.py
@@ -10,15 +10,15 @@ from markdown_it.tree import SyntaxTreeNode
 
 def get_issue(issue_id: str) -> dict[str, str]:
     """Returns the issue with the given id"""
-    url = f"https://api.github.com/repos/kjaymiller/Python-Community-News/issues/{issue_id}"  # TODO: remove hardcoded issue url
+    url = f"https://api.github.com/repos/kjaymiller/Python-Community-News/issues/{issue_id}"
     request = httpx.get(url)
     return request.json()
 
 
-def get_issues(labels, since_date: datetime.datetime | None) -> list:
+def get_issues(labels, **kwargs) -> list:
     """Returns the issues filed in the last week"""
     url = "https://api.github.com/repos/kjaymiller/Python-Community-News/issues"
-    params = {"labels": ",".join(labels), "since": since_date}
+    params = {**{"labels": ",".join(labels)}, **kwargs}
     request = httpx.get(url, params=params)
     return request.json()
