+++ automation/src/newsletter.py
